requirements.txt
aiohttp
pydantic
python-dotenv
SQLAlchemy
aiosqlite